Moved tests to CheckoutOrderTest
