+++ modules/checkout/tests/src/Functional/CheckoutOrderTest.php
@@ -8,7 +8,7 @@ use Drupal\Tests\commerce\Functional\CommerceBrowserTestBase;
 /**
  * Tests the checkout of an order.
  *
- * @group commerce
+ * @group niels
  */
 class CheckoutOrderTest extends CommerceBrowserTestBase {
 
@@ -36,7 +36,7 @@ class CheckoutOrderTest extends CommerceBrowserTestBase {
   public static $modules = ['system', 'field', 'user', 'text',
     'entity', 'views', 'address', 'profile', 'commerce', 'inline_entity_form',
     'commerce_price', 'commerce_store', 'commerce_product', 'commerce_cart',
-    'commerce_checkout', 'commerce_order',
+    'commerce_checkout', 'commerce_order', 'views_ui',
   ];
 
   /**
@@ -66,6 +66,16 @@ class CheckoutOrderTest extends CommerceBrowserTestBase {
   }
 
   /**
+   * {@inheritdoc}
+   */
+  protected function getAdministratorPermissions() {
+    return array_merge([
+      'administer checkout flows',
+      'administer views',
+    ], parent::getAdministratorPermissions());
+  }
+
+  /**
    * Tests than an order can go through checkout steps.
    */
   public function testGuestOrderCheckout() {
@@ -75,6 +85,7 @@ class CheckoutOrderTest extends CommerceBrowserTestBase {
     $cart_link = $this->getSession()->getPage()->findLink('your cart');
     $cart_link->click();
     $this->submitForm([], 'Checkout');
+    $this->assertSession()->pageTextNotContains('Order Summary');
     $this->submitForm([], 'Continue as Guest');
     $this->submitForm([
       'contact_information[email]' => 'guest@example.com',
@@ -86,8 +97,39 @@ class CheckoutOrderTest extends CommerceBrowserTestBase {
     ], 'Continue to review');
     $this->assertSession()->pageTextContains('Contact information');
     $this->assertSession()->pageTextContains('Billing information');
+    $this->assertSession()->pageTextContains('Order Summary');
     $this->submitForm([], 'Pay and complete purchase');
     $this->assertSession()->pageTextContains('Your order number is 1. You can view your order on your account page when logged in.');
+  }
+
+  /**
+   * Test OrderSummary settings.
+   */
+  public function testOrderSummarySettings() {
+    // Fill the cart.
+    $this->drupalGet($this->product->toUrl()->toString());
+    $this->submitForm([], 'Add to cart');
+
+    // Test OrderSummary default behavior.
+    $this->drupalGet('/checkout/1');
+    // Validate that the default OrderSummary-View is on the checkout-page.
+    $this->assertSession()->elementExists('css', '.view-id-commerce_checkout_order_summary');
+
+    // Disable OrderSummary.
+    $this->drupalGet('/admin/commerce/config/checkout-flows/manage/default');
+    $this->submitForm(['configuration[order_summary_view]' => ''], t('Save'));
+    $this->drupalGet('/checkout/1');
+    // Validate that the default OrderSummary-View is not on the checkout-page.
+    $this->assertSession()->elementNotExists('css', '.view-id-commerce_checkout_order_summary');
+
+    // Change OrderSummary to a different View.
+    $this->drupalGet('/admin/structure/views/view/commerce_checkout_order_summary/duplicate');
+    $this->submitForm(['id' => 'duplicate_of_commerce_checkout_order_summary'], 'Duplicate');
+    $this->drupalGet('/admin/commerce/config/checkout-flows/manage/default');
+    $this->submitForm(['configuration[order_summary_view]' => 'duplicate_of_commerce_checkout_order_summary'], t('Save'));
+    $this->drupalGet('/checkout/1');
+    // Validate that the duplicated OrderSummary-View is on the checkout-page.
+    $this->assertSession()->elementExists('css', '.view-id-duplicate_of_commerce_checkout_order_summary');
   }
 
 }
